% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNAs.R
\name{plotNAs}
\alias{plotNAs}
\title{Plot the Percenage of NAs for each Variables}
\usage{
plotNAs(df)
}
\arguments{
\item{df}{This is a dataframe}
}
\value{
A bar chart that shows the Percenage of NAs for each Variables
}
\description{
Plot the Percenage of NAs for each Variables
}
\examples{
plotNAs(ucla)
plotNAs(df)


}
