% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\docType{data}
\name{store}
\alias{store}
\title{Web Scraped Data of Google Play Store Apps}
\format{
10,841 X 12 dataframe
}
\usage{
store
}
\description{
This is a Web scraped dataset with 9k+ Play Store apps for analysing the Android market. Each app (row) has values for catergory, rating, size, and more.
}
\keyword{datasets}
